requirements.txt
Adafruit-Blinka==8.19.0
adafruit-circuitpython-bme680==3.5.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==1.14.1
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.46.0
Adafruit-PureIO==1.1.11
board==1.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pms5003==0.0.5
pyftdi==0.54.0
pyserial==3.5
python-dotenv==1.0.0
pyusb==1.2.1
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing-extensions==4.7.1
urllib3==2.0.3
